Rocket telemetry producer/requirements.txt
quixstreams==0.5.7
python-dotenv==0.10.1